+++ dwds/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 24.4.0+1-wip
+
+### Bug Fixes:
+
+- Fix issue where DDS would fail to initialize if DWDS already had existing
+  clients.
+
 ## 24.4.0
 
 - Added support for breakpoint registering on a hot reload with the DDC library bundle format using PausePostRequests.
@@ -42,6 +49,7 @@
   libraries instead of assuming it exists in the global `window`.
 
 ## 24.3.5
+
 - Allow clients to specify the `packageConfigPath` in `LoadStrategy` class and associated providers.
 
 ## 24.3.4
@@ -126,7 +134,7 @@
 
 ## 23.1.1
 
-- Loosen `package:vm_service` constraints to allow `>=13.0.0 <15.0.0`.  - [#2329](https://github.com/dart-lang/webdev/pull/2329)
+- Loosen `package:vm_service` constraints to allow `>=13.0.0 <15.0.0`. - [#2329](https://github.com/dart-lang/webdev/pull/2329)
 
 ## 23.1.0